Update main.c
changed type `int` to `size_t`

possible overflow issue when `strlen(str) + 1 > INT_MAX`

+++ main.c
@@ -62,7 +62,7 @@
  */
 char *strdup(const char *str)
 {
-    int n = strlen(str) + 1;
+    size_t n = strlen(str) + 1;
     char *dup = malloc(n);
     if(dup)
     {
